Correction of an Issue
The issue pertained to the abnormal usage of the r2_score function. Typically, the function is employed as r2_score(y_true, y_predicted); however, in one instance, it was mistakenly used as r2_score(y_predicted, y_true). Despite this error, the final results appear to be very similar.